riportate modifiche per time management

+++ thread.h
@@ -32,9 +32,7 @@
 
 struct timeManagementStruct
 {
-	volatile unsigned long allocatedTime;
-	//volatile unsigned long minSearchTime;
-	//volatile unsigned long maxSearchTime;
+	volatile long long allocatedTime;
 	volatile unsigned int depth;
 	volatile unsigned int singularRootMoveCount;
 	volatile unsigned int resolution;
